backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.10
django-cors-headers==4.2.0
Pillow==11.2.1
celery==5.3.4
redis==5.0.1
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
django-filter==23.2
django-storages==1.14.2
social-auth-app-django==5.2.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
django-jazzmin==2.6.2
python-slugify==8.0.4